run as roor

+++ build.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ `whoami` != 'root' ]]; then
+  echo 'Run this script as root!' 1>&2
+  exit 1
+fi
+
 # remove
 apt-get -yq remove audacious                                                                           \
                    ace-of-penguins                                                                     \
